==> notebooks/08.ridge-regression.ipynb <==
import pandas as pd
import numpy as np
import pickle

from sklearn.linear_model import RidgeCV
from helpers import target_feature, random_seed


np.random.seed(random_seed)

## Ridge regression

train = pd.read_csv('../data/post_fs_train.csv', index_col=0)
test = pd.read_csv('../data/post_fs_test.csv', index_col=0)

y_train = train[target_feature]
y_test = test[target_feature]

x_train = train.drop(target_feature, axis=1)
x_test = test.drop(target_feature, axis=1)

compute_cv = True

if compute_cv:
    model = RidgeCV(cv=5, alphas=[i*0.025 for i in range(1, 51)]).fit(x_train.values, y_train)
    score = model.score(x_train.values, y_train)
    print(score, model.alpha_)

    with open('../models/ridge.pkl', 'wb') as f:
        pickle.dump(model, f) 

with open('../models/ridge.pkl', 'rb') as f:
    ridge = pickle.load(f)

ridge.alpha_
